+++ packages/status-js/src/index.ts
@@ -1,6 +1,6 @@
 export type { Account } from './account'
-export type { Client, ClientOptions } from './client'
-export { createClient } from './client'
 export type { ChatMessage as Message } from './client/chat'
+export type { Client, ClientOptions } from './client/client'
+export { createClient } from './client/client'
 export type { Community } from './client/community/community'
 export type { Member } from './client/member'
